vercel/requirements.txt
pandas
numpy
requests
beautifulsoup4
polygon-api-client
tweepy
psutil
matplotlib
python-dotenv
